1.1
Added .gitignore for backend bin
Added styling
The rectangle can be moved, and resized in any way.
X and Y coordinates of the rectangle are tracked and displayed.
The rectangle now cannot go through the svg border.
Moved the verification complete popup to the bottom of the panel instead of it being inside the rectangle.